requirements.txt
Flask==2.3.3
gunicorn==20.1.0
flask-cors
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
supabase==2.3.4
postgrest==0.13.2